requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
binaryornot==0.4.4
boto3==1.35.21
botocore==1.35.21
celery==5.4.0
celery-progress==0.4
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
crispy-tailwind==1.0.3
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.1
django-allauth==64.1.0
django-browser-reload==1.13.0
django-celery-results==2.5.1
django-ckeditor-5==0.2.14
django-cleanup==8.1.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-import-export==4.1.1
django-pivot==1.10.0
django-recaptcha==4.0.0
django-storages==1.14.4
django-tailwind==3.8.0
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
flower==2.0.1
humanize==4.10.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.1
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.1
odfpy==1.4.1
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
rich==13.7.1
ruff==0.6.2
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
tabulate==0.9.0
text-unidecode==1.3
tornado==6.4.1
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
xlrd==2.0.1
xlwt==1.3.0
